activity-editor-descriptionを追加
