+++ assets/locales/components/organisms/activity-editor-description.json
@@ -1,0 +1,8 @@
+{
+  "en": {
+    "description": "Description"
+  },
+  "ja": {
+    "description": "作業内容や補足を入力"
+  }
+}
