+++ components/organisms/activity-editor.vue
@@ -7,31 +7,10 @@
     </modal-header>
 
     <form @submit.prevent="saveActivity">
-      <modal-item>
-        <input
-          v-model="description"
-          :placeholder="$t('description')"
-          type="text"
-          @focus="focus"
-          @blur="blur"
-          @input="input"
-        />
-      </modal-item>
-
-      <transition>
-        <div v-if="focused" class="suggestion-list">
-          <ul>
-            <li
-              v-for="(suggestion, index) in suggestions"
-              :key="index"
-              class="suggestion"
-              @click="clickSuggestion(suggestion)"
-            >
-              <activity-name v-bind="suggestion" />
-            </li>
-          </ul>
-        </div>
-      </transition>
+      <activity-editor-description
+        :description.sync="description"
+        :project.sync="project"
+      />
 
       <modal-item>
         <button type="button" class="project-button" @click="editProject">
@@ -86,7 +65,7 @@
 </template>
 
 <script>
-import ActivityName from '@/components/molecules/activity-name';
+import ActivityEditorDescription from '@/components/organisms/activity-editor-description';
 import ProjectList from '@/components/organisms/project-list';
 import ModalItem from '@/components/molecules/modal-item';
 import ModalHeader from '@/components/molecules/modal-header';
@@ -97,13 +76,11 @@ import BaseButton from '@/components/atoms/base-button';
 import Icon from '@/components/atoms/icon';
 import { formatDistanceStrict, parseISO } from 'date-fns';
 import { mapGetters } from 'vuex';
-import { disableBodyScroll, enableBodyScroll } from 'body-scroll-lock';
-import debounce from 'lodash.debounce';
 
 export default {
   name: 'ActivityEditor',
   components: {
-    ActivityName,
+    ActivityEditorDescription,
     ProjectName,
     DatetimePicker,
     ModalHeader,
@@ -221,27 +198,6 @@ export default {
         if (e.name === 'AbortError') return;
         throw e;
       }
-    },
-    fetchSuggestions: debounce(function(text) {
-      this.$store.dispatch('suggestions/fetch', text);
-    }, 1000),
-    input(e) {
-      this.description = e.target.value;
-      this.fetchSuggestions(this.description);
-    },
-    focus(e) {
-      this.focused = true;
-      e.target.select();
-      this.fetchSuggestions('');
-      disableBodyScroll(e.target);
-    },
-    blur(e) {
-      this.focused = false;
-      enableBodyScroll(e.target);
-    },
-    clickSuggestion({ description, project }) {
-      this.description = description;
-      this.project = project;
     }
   }
 };
@@ -258,35 +214,6 @@ export default {
   height: 40px;
   width: 100%;
   background: none;
-}
-.suggestion-list {
-  height: 310px;
-  overflow: hidden;
-  overflow: scroll;
-  position: absolute;
-  animation-duration: 100ms;
-  width: 100%;
-  box-shadow: 0 8px 5px -5px $shadow-dark inset;
-  background-color: $background;
-}
-.suggestion-list ul {
-  width: 100%;
-  list-style-type: none;
-  list-style-position: inside;
-  padding: 0;
-  margin: 0;
-  padding-bottom: 200px;
-}
-.suggestion-list li {
-  display: flex;
-  padding: 0 30px;
-  height: 65px;
-  align-items: center;
-  border-bottom: 1px solid $border;
-  transition: background-color 0.1s ease;
-  &:hover {
-    background-color: $background-hover;
-  }
 }
 .icons {
   display: flex;
